# docs/reference/api/openapi.yaml
openapi: 3.0.3
info:
  title: Family-Tree API
  version: 0.1.0
  description: |
    Draft OpenAPI based on current guides.
    Security: bearer access token (JWT). Refresh via httpOnly cookie.
servers:
  - url: http://localhost:8080
...
paths:
  /auth/telegram/session:
    post:
      tags: [Auth]
      summary: Create pending session and return deep-link
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTelegramSessionRequest'
      responses:
        '200':
          description: Created pending session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateTelegramSessionResponse'
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /auth/telegram/session/{sid}:
    get:
      tags: [Auth]
      summary: Poll pending session status or receive AuthResponse
      parameters:
        - name: sid
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Status or AuthResponse
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/PendingSessionStatus'
                  - $ref: '#/components/schemas/PendingSessionReady'
        '404': { $ref: '#/components/responses/NotFound' }
        '410': { description: Gone }
        '409': { $ref: '#/components/responses/Conflict' }
  /auth/telegram/webhook:
    post:
      tags: [Auth]
      summary: Telegram webhook endpoint
      description: Requires header `X-Telegram-Bot-Api-Secret-Token` matching configured secret.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelegramUpdate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  message: { type: string }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: refreshToken
...
  responses:
...
    NotFound:
      description: Not Found
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Conflict:
      description: Conflict
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    TooManyRequests:
      description: Too Many Requests
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
    TelegramUpdate:
      type: object
      properties:
        update_id: { type: integer, format: int64 }
        message:
          type: object
          properties:
            message_id: { type: integer, format: int64 }
            text: { type: string }
            from:
              type: object
              properties:
                id: { type: integer, format: int64 }
                username: { type: string }
            chat:
              type: object
              properties:
                id: { type: integer, format: int64 }
    Error:
      type: object
      properties:
        error: { type: string }
        message: { type: string }
    CreateTelegramSessionRequest:
      type: object
      properties:
        invitationId: { type: string }
    CreateTelegramSessionResponse:
      type: object
      properties:
        sid: { type: string }
        deeplinkUrl: { type: string }
        expiresIn: { type: integer }
      required: [sid, deeplinkUrl, expiresIn]
    PendingSessionStatus:
      type: object
      properties:
        status: { type: string, enum: [pending] }
    PendingSessionReady:
      type: object
      properties:
        status: { type: string, enum: [ready] }
        auth: { $ref: '#/components/schemas/AuthResponse' }
      required: [status, auth]
    AuthResponse:
      type: object
      properties:
        accessToken: { type: string }
        expiresIn: { type: integer }
        user:
          type: object
          properties:
            id: { type: string }
            name: { type: string }
        activeFamilyId: { type: string }
        scopes:
          type: array
          items: { type: string }
        refreshTokenSet: { type: boolean }
      required: [accessToken, expiresIn]
...